+++ CHANGELOG.md
@@ -2,6 +2,10 @@
 All notable changes to this project will be documented in this file.
 This project adheres to [Semantic Versioning](http://semver.org/).
 
+## [Unreleased]
+### Fixed
+- Improve `repr` implementation for SQL type wrapper objects.
+
 ## [1.2.2] - 2017-01-10
 ### Fixed
 - Fix multi-byte UTF-16 character conversion.